Extract core datasources to top-level classes.
