+++ src/commonTest/kotlin/baaahs/gl/patch/AutoWirerSpec.kt
@@ -10,8 +10,11 @@ import baaahs.gl.shader.OpenShader
 import baaahs.gl.shader.OutputPort
 import baaahs.glsl.Shaders.cylindricalProjection
 import baaahs.only
-import baaahs.plugin.CorePlugin
 import baaahs.plugin.core.FixtureInfoDataSource
+import baaahs.plugin.core.datasource.ModelInfoDataSource
+import baaahs.plugin.core.datasource.ResolutionDataSource
+import baaahs.plugin.core.datasource.SliderDataSource
+import baaahs.plugin.core.datasource.TimeDataSource
 import baaahs.show.Shader
 import baaahs.show.ShaderChannel
 import baaahs.show.live.FakeOpenShader
@@ -84,7 +87,7 @@ object AutoWirerSpec : Spek({
                     override(portContentType) { ContentType.Time }
 
                     it("suggests the data source's channel") {
-                        expect(portLink).toBe(CorePlugin.TimeDataSource().editor())
+                        expect(portLink).toBe(TimeDataSource().editor())
                     }
 
                     // TODO: get these working?
@@ -92,14 +95,14 @@ object AutoWirerSpec : Spek({
                         override(outContentType) { portContentType }
 
                         it("suggests the data source's channel") {
-                            expect(portLink).toBe(CorePlugin.TimeDataSource().editor())
+                            expect(portLink).toBe(TimeDataSource().editor())
                         }
 
                         context("and the shader's channel matches the data source's channel") {
                             override(shaderChannel) { ShaderChannel("time") }
 
                             it("suggests the data source's channel") {
-                                expect(portLink).toBe(CorePlugin.TimeDataSource().editor())
+                                expect(portLink).toBe(TimeDataSource().editor())
                             }
                         }
                     }
@@ -111,7 +114,7 @@ object AutoWirerSpec : Spek({
 
                     it("suggests a Slider data source channel link") {
                         expect(portLink)
-                            .toBe(CorePlugin.SliderDataSource("Brightness", 1f, 0f, 1f).editor())
+                            .toBe(SliderDataSource("Brightness", 1f, 0f, 1f).editor())
                     }
                 }
 
@@ -191,24 +194,24 @@ object AutoWirerSpec : Spek({
 
             it("picks TimeDataSource for time") {
                 expect(mutableLinks["time"])
-                    .toBe(CorePlugin.TimeDataSource().editor())
+                    .toBe(TimeDataSource().editor())
 
                 expect(links["time"])
-                    .toBe(CorePlugin.TimeDataSource().link("time"))
+                    .toBe(TimeDataSource().link("time"))
             }
 
             it("picks ResolutionDataSource for resolution") {
                 expect(mutableLinks["resolution"])
-                    .toBe(CorePlugin.ResolutionDataSource().editor())
+                    .toBe(ResolutionDataSource().editor())
 
                 expect(links["resolution"])
-                    .toBe(CorePlugin.ResolutionDataSource().link("resolution"))
+                    .toBe(ResolutionDataSource().link("resolution"))
             }
 
             it("picks a Slider for blueness") {
                 expect(links["blueness"])
                     .toBe(
-                        CorePlugin.SliderDataSource("Blueness", 1f, 0f, 1f, null)
+                        SliderDataSource("Blueness", 1f, 0f, 1f, null)
                             .link("bluenessSlider")
                     )
             }
@@ -226,9 +229,9 @@ object AutoWirerSpec : Spek({
                     .toBe(
                         mapOf(
                             "gl_FragCoord" to DefaultValueNode(ContentType.UvCoordinate),
-                            "time" to CorePlugin.TimeDataSource().link("time"),
-                            "resolution" to CorePlugin.ResolutionDataSource().link("resolution"),
-                            "blueness" to CorePlugin.SliderDataSource("Blueness", 1f, 0f, 1f, null)
+                            "time" to TimeDataSource().link("time"),
+                            "resolution" to ResolutionDataSource().link("resolution"),
+                            "blueness" to SliderDataSource("Blueness", 1f, 0f, 1f, null)
                                 .link("bluenessSlider")
                         )
                     )
@@ -255,7 +258,7 @@ object AutoWirerSpec : Spek({
                                 MutableShader(mainShader),
                                 hashMapOf(
                                     "fragCoord" to ShaderChannel.Main.editor(),
-                                    "resolution" to CorePlugin.ResolutionDataSource().editor(),
+                                    "resolution" to ResolutionDataSource().editor(),
                                 ),
                                 shaderChannel = ShaderChannel.Main.editor(),
                                 priority = 0f
@@ -292,9 +295,9 @@ object AutoWirerSpec : Spek({
                             MutableShaderInstance(
                                 MutableShader(mainShader),
                                 hashMapOf(
-                                    "iTime" to CorePlugin.TimeDataSource().editor(),
-                                    "blueness" to CorePlugin.SliderDataSource("Blueness", 1f, 0f, 1f, null).editor(),
-                                    "iResolution" to CorePlugin.ResolutionDataSource().editor(),
+                                    "iTime" to TimeDataSource().editor(),
+                                    "blueness" to SliderDataSource("Blueness", 1f, 0f, 1f, null).editor(),
+                                    "iResolution" to ResolutionDataSource().editor(),
                                     "fragCoord" to ShaderChannel.Main.editor()
                                 ),
                                 shaderChannel = ShaderChannel.Main.editor(),
@@ -310,10 +313,10 @@ object AutoWirerSpec : Spek({
                     }
                     expects(
                         mapOf(
-                            "blueness" to CorePlugin.SliderDataSource("Blueness", 1f, 0f, 1f, null)
+                            "blueness" to SliderDataSource("Blueness", 1f, 0f, 1f, null)
                                 .link("bluenessSlider"),
-                            "iResolution" to CorePlugin.ResolutionDataSource().link("resolution"),
-                            "iTime" to CorePlugin.TimeDataSource().link("time"),
+                            "iResolution" to ResolutionDataSource().link("resolution"),
+                            "iTime" to TimeDataSource().link("time"),
                             "fragCoord" to DefaultValueNode(ContentType.UvCoordinate)
                         )
                     ) { rootProgramNode.incomingLinks }
@@ -332,9 +335,9 @@ object AutoWirerSpec : Spek({
                             MutableShaderInstance(
                                 MutableShader(mainShader),
                                 hashMapOf(
-                                    "time" to CorePlugin.TimeDataSource().editor(),
-                                    "resolution" to CorePlugin.ResolutionDataSource().editor(),
-                                    "blueness" to CorePlugin.SliderDataSource("Blueness", 1f, 0f, 1f, null).editor(),
+                                    "time" to TimeDataSource().editor(),
+                                    "resolution" to ResolutionDataSource().editor(),
+                                    "blueness" to SliderDataSource("Blueness", 1f, 0f, 1f, null).editor(),
                                     "gl_FragCoord" to ShaderChannel.Main.editor()
                                 ),
                                 shaderChannel = ShaderChannel.Main.editor()
@@ -342,7 +345,7 @@ object AutoWirerSpec : Spek({
                             uvShaderInst.apply {
                                 incomingLinks.putAll(
                                     mapOf(
-                                        "modelInfo" to CorePlugin.ModelInfoDataSource().editor(),
+                                        "modelInfo" to ModelInfoDataSource().editor(),
                                         "pixelLocation" to PixelLocationDataSource().editor()
                                     )
                                 )
@@ -376,7 +379,7 @@ object AutoWirerSpec : Spek({
                             MutableShaderInstance(
                                 MutableShader(filterShader),
                                 hashMapOf(
-                                    "brightness" to CorePlugin.SliderDataSource("Brightness", 1f, 0f, 1f, null)
+                                    "brightness" to SliderDataSource("Brightness", 1f, 0f, 1f, null)
                                         .editor(),
                                     "colorIn" to ShaderChannel.Main.editor()
                                 ),
