+++ src/commonTest/kotlin/baaahs/gl/render/RenderEngineTest.kt
@@ -12,7 +12,7 @@ import baaahs.geom.Vector3F
 import baaahs.gl.*
 import baaahs.gl.glsl.GlslProgram
 import baaahs.gl.patch.ContentType
-import baaahs.plugin.CorePlugin
+import baaahs.plugin.core.datasource.GadgetDataSource
 import baaahs.show.Shader
 import baaahs.shows.FakeShowPlayer
 import kotlin.math.abs
@@ -257,7 +257,7 @@ class RenderEngineTest {
             .confirm()
             .openForPreview(testToolchain, ContentType.Color)!!
         return renderEngine.compile(linkedPatch) { id, dataSource ->
-            if (dataSource is CorePlugin.GadgetDataSource<*>) {
+            if (dataSource is GadgetDataSource<*>) {
                 fakeShowPlayer.registerGadget(id, dataSource.createGadget(), dataSource)
             }
             dataSource.createFeed(fakeShowPlayer, id)
