+++ src/commonTest/kotlin/baaahs/plugin/CorePluginSpec.kt
@@ -9,6 +9,7 @@ import baaahs.gl.patch.ContentType
 import baaahs.gl.testPlugins
 import baaahs.glsl.Shaders
 import baaahs.only
+import baaahs.plugin.core.datasource.TimeDataSource
 import baaahs.shows.FakeGlContext
 import baaahs.shows.FakeShowPlayer
 import ch.tutteli.atrium.api.fluent.en_GB.toBe
@@ -16,9 +17,9 @@ import ch.tutteli.atrium.api.verbs.expect
 import org.spekframework.spek2.Spek
 
 object CorePluginSpec : Spek({
-    describe<CorePlugin.TimeDataSource> {
+    describe<TimeDataSource> {
         val clock by value { FakeClock(0.0) }
-        val dataSource by value { CorePlugin.TimeDataSource() }
+        val dataSource by value { TimeDataSource() }
         val toolchain by value { RootToolchain(testPlugins(clock) ) }
         val feed by value { dataSource.createFeed(FakeShowPlayer(toolchain = toolchain), "time") }
         val gl by value { FakeGlContext() }
